Give pairing genomes the ability to calculate their own fitness.
